Edited home page header, details, plant card, and data files
