--- lib/models/myPlant.dart
@@ -1,10 +1,0 @@
-class MyPlant {
-  String plantName, details, imgURL, state;
-
-  MyPlant({
-    this.imgURL,
-    this.plantName,
-    this.details,
-    this.state,
-  });
-}
