--- lib/models/plant.dart
@@ -1,7 +1,0 @@
-class Plant {
-  String plantName, imgURL;
-  Plant({
-    this.plantName,
-    this.imgURL,
-  });
-}
