+++ lib/screens/home/components/header.dart
@@ -20,11 +20,11 @@ class HeaderWithSearchBox extends StatelessWidget {
         children: <Widget>[
           Container(
             padding: EdgeInsets.only(
-              left: kDefaultPadding,
+              left: 0,
               right: kDefaultPadding,
               bottom: 36 + kDefaultPadding,
             ),
-            height: size.height * 0.25 - 27,
+            height: size.height * 0.25 - 25,
             width: size.width,
             decoration: BoxDecoration(
               color: kPrimaryColor,
@@ -35,14 +35,26 @@ class HeaderWithSearchBox extends StatelessWidget {
             ),
             child: Column(
               children: <Widget>[
-                Text(
-                  'Explore',
-                  style: Theme.of(context).textTheme.headline5.copyWith(
-                      color: Colors.white, fontWeight: FontWeight.bold),
-                ),
-                Text(
-                  'Add new plants to your family',
-                  style: TextStyle(color: Colors.white, fontSize: 12),
+                RichText(
+                  text: TextSpan(
+                    children: [
+                      TextSpan(
+                        text: 'Explore\n',
+                        style: Theme.of(context).textTheme.headline5.copyWith(
+                              color: Colors.white,
+                              fontWeight: FontWeight.bold,
+                              letterSpacing: 1.5,
+                            ),
+                      ),
+                      TextSpan(
+                        text: 'Add new plants to your family',
+                        style: TextStyle(
+                            color: Colors.white,
+                            fontSize: 14,
+                            letterSpacing: 0.5),
+                      ),
+                    ],
+                  ),
                 ),
               ],
             ),
@@ -53,9 +65,9 @@ class HeaderWithSearchBox extends StatelessWidget {
             right: 0,
             child: Container(
               alignment: Alignment.center,
-              margin: EdgeInsets.symmetric(horizontal: kDefaultPadding),
+              margin: EdgeInsets.symmetric(horizontal: kDefaultPadding + 5),
               padding: EdgeInsets.symmetric(horizontal: kDefaultPadding),
-              height: 54,
+              height: 43,
               decoration: BoxDecoration(
                 color: Colors.white,
                 borderRadius: BorderRadius.circular(20),
